MB-3398 Move check_response to tasks/base.py
